// packages/adk/src/namespaces/atom/types.ts
export type AtomRelation =
  | 'http://www.sap.com/adt/relations/versions'
  | 'http://www.sap.com/adt/relations/source'
  | 'http://www.sap.com/adt/relations/objectstructure'
  | 'http://www.sap.com/adt/relations/sources/withabapdocfromshorttexts'
  | 'http://www.sap.com/adt/relations/objectstates'
  | 'http://www.sap.com/adt/relations/abapsource/parser';